+++ R/tokenize.R
@@ -1,3 +1,4 @@
+utils::globalVariables(c("smart_stopwords"))
 #' Tokenize a character vector
 
 #' Parse the elements of a character vector into a list of cleaned tokens.
@@ -6,11 +7,10 @@
 #' @param removeNum \code{TRUE} or \code{FALSE} indicating whether or not to remove numbers from \code{text}.  If \code{TRUE}, numbers will be removed.  Defaults to \code{TRUE}.
 #' @param toLower \code{TRUE} or \code{FALSE} indicating whether or not to coerce all of \code{text} to lowercase.  If \code{TRUE}, \code{text} will be coerced to lowercase.  Defaults to \code{TRUE}.
 #' @param stemWords \code{TRUE} or \code{FALSE} indicating whether or not to stem resulting tokens.  If \code{TRUE}, the outputted tokens will be tokenized using \code{SnowballC::wordStem()}.  Defaults to \code{TRUE}.
-#' @param rmStopWords \code{TRUE}, \code{FALSE}, or character vector of stopwords to remove. If \code{TRUE}, words in \code{tm::stopwords("SMART")} will be removed prior to stemming. If \code{FALSE}, no stopword removal will occur. If a character vector is passed, this vector will be used as the list of stopwords to be removed.  Defaults to \code{TRUE}.
+#' @param rmStopWords \code{TRUE}, \code{FALSE}, or character vector of stopwords to remove. If \code{TRUE}, words in \code{lexRankr::smart_stopwords} will be removed prior to stemming. If \code{FALSE}, no stopword removal will occur. If a character vector is passed, this vector will be used as the list of stopwords to be removed.  Defaults to \code{TRUE}.
 #' @examples
 #' tokenize("Mr. Feeny said the test would be on Sat. At least I'm 99.9% sure that's what he said.")
 #' tokenize("Bill is trying to earn a Ph.D. in his field.", rmStopWords=FALSE)
-#' @importFrom magrittr "%>%"
 
 #' @export
 tokenize <- function(text, removePunc=TRUE, removeNum=TRUE, toLower=TRUE, stemWords=TRUE, rmStopWords=TRUE){
@@ -32,7 +32,7 @@ tokenize <- function(text, removePunc=TRUE, removeNum=TRUE, toLower=TRUE, stemWo
     if(length(rmStopWords) != 1) stop("rmStopWords must be length 1 if passed as a logical")
     if(rmStopWords) {
       rmStopWordFlag <- TRUE
-      stopwords <- tm::stopwords("SMART")
+      stopwords <- smart_stopwords
     } else {
       rmStopWordFlag <- FALSE
     }
@@ -42,11 +42,9 @@ tokenize <- function(text, removePunc=TRUE, removeNum=TRUE, toLower=TRUE, stemWo
   if (removeNum) text <- gsub(x=text,pattern="([[:digit:]])",replacement="")
   if (toLower) text <- tolower(text)
 
-  text <- text %>%
-    gsub(pattern="([^[:alnum:] ])",replacement=" \\1 ") %>% 
-    gsub(pattern="\\s+",replacement=" ") %>%
-    trimws() %>% 
-    stringr::str_split(pattern=" ")
+  text <- gsub(x=text, pattern="([^[:alnum:] ])",replacement=" \\1 ")
+  text <- trimws(gsub(x=text, pattern="\\s+",replacement=" "))
+  text <- strsplit(x=text, split=" ", fixed=TRUE)
 
   if(rmStopWordFlag) text <- lapply(text, function(tokens) {
     checkTokens <- tolower(tokens)
@@ -58,7 +56,12 @@ tokenize <- function(text, removePunc=TRUE, removeNum=TRUE, toLower=TRUE, stemWo
     if(length(nonStopTok) == 0) NA_character_ else nonStopTok
   })
   if(stemWords) {
-    text <- lapply(text, SnowballC::wordStem)
+    text <- lapply(text, function(w) {
+      w_na = which(is.na(w))
+      out = SnowballC::wordStem(w)
+      out[w_na] = NA
+      out
+    })
   }
 
   tokenList <- lapply(text, function(tokens) {