+++ tests/testthat/test-sentenceSimil.R
@@ -9,7 +9,32 @@ test_that("testing result str and class", {
   
   testResult <- sentenceSimil(sentenceId = tokenDf$sentenceId,
                               token = tokenDf$token,
-                              docId = tokenDf$docId)
+                              docId = tokenDf$docId,
+                              sentencesAsDocs = FALSE)
+  
+  expect_equal(class(testResult), "data.frame")
+  expect_equal(names(testResult), c("sent1","sent2","similVal"))
+  
+  expect_true(is.character(testResult$sent1))
+  expect_true(is.character(testResult$sent2))
+  expect_true(is.numeric(testResult$similVal))
+  
+  testResult <- sentenceSimil(sentenceId = tokenDf$sentenceId,
+                              token = tokenDf$token,
+                              docId = tokenDf$docId,
+                              sentencesAsDocs = TRUE)
+  
+  expect_equal(class(testResult), "data.frame")
+  expect_equal(names(testResult), c("sent1","sent2","similVal"))
+  
+  expect_true(is.character(testResult$sent1))
+  expect_true(is.character(testResult$sent2))
+  expect_true(is.numeric(testResult$similVal))
+  
+  testResult <- sentenceSimil(sentenceId = tokenDf$sentenceId,
+                              token = tokenDf$token,
+                              docId = tokenDf$docId,
+                              sentencesAsDocs = TRUE)
   
   expect_equal(class(testResult), "data.frame")
   expect_equal(names(testResult), c("sent1","sent2","similVal"))
@@ -18,7 +43,7 @@ test_that("testing result str and class", {
   expect_true(is.character(testResult$sent2))
   expect_true(is.numeric(testResult$similVal))
 })
-  
+
 
 test_that("bad input", {
   expect_error(sentenceSimil(sentenceId = c("1_1"),
@@ -29,12 +54,13 @@ test_that("bad input", {
                              token = c(1,2),
                              docId = c(1,2)))
   
-  testDocs <- c("test","test")
-  tokenDf <- sentenceTokenParse(testDocs)$tokens
-  
-  expect_error(sentenceSimil(sentenceId = tokenDf$sentenceId,
-                             token = tokenDf$token,
-                             docId = tokenDf$docId))
+  #was relevant when using idf calc w/o bounding at 1
+  # testDocs <- c("test","test")
+  # tokenDf <- sentenceTokenParse(testDocs)$tokens
+  # 
+  # expect_error(sentenceSimil(sentenceId = tokenDf$sentenceId,
+  #                            token = tokenDf$token,
+  #                            docId = tokenDf$docId))
   
   testDocs <- c("1","2")
   tokenDf <- sentenceTokenParse(testDocs)$tokens
@@ -54,12 +80,26 @@ test_that("output value check", {
   
   testResult <- sentenceSimil(sentenceId = tokenDf$sentenceId,
                               token = tokenDf$token,
-                              docId = tokenDf$docId) %>% 
-    dplyr::mutate(similVal = round(similVal, 5))
+                              docId = tokenDf$docId,
+                              sentencesAsDocs = FALSE)
+  testResult$similVal = round(testResult$similVal, 5)
   
   expectedResult <- data.frame(sent1 = c("1_1", "1_1", "2_1"),
                                sent2 = c("2_1", "3_1", "3_1"),
-                               similVal = c(0.32718, 0, 0.32718),
+                               similVal = c(0.48624, 0, 0.48624),
+                               stringsAsFactors = FALSE)
+  
+  expect_equal(testResult, expectedResult)
+  
+  testResult <- sentenceSimil(sentenceId = tokenDf$sentenceId,
+                              token = tokenDf$token,
+                              docId = tokenDf$docId,
+                              sentencesAsDocs = TRUE)
+  testResult$similVal = round(testResult$similVal, 5)
+  
+  expectedResult <- data.frame(sent1 = c("1_1", "1_1", "2_1"),
+                               sent2 = c("2_1", "3_1", "3_1"),
+                               similVal = c(0.48624, 0, 0.48624),
                                stringsAsFactors = FALSE)
   
   expect_equal(testResult, expectedResult)